ir: add stream ID API
The new bt_ctf_stream_create_with_id() function creates a stream object,
just like bt_ctf_stream_create(), but with a numeric ID. You cannot
call this function to create a CTF writer stream because CTF writer
streams initialize their own unique ID.

Use bt_ctf_stream_get_id() to get a stream's ID.

+++ include/babeltrace/ctf-ir/stream-internal.h
@@ -49,7 +49,7 @@ struct bt_ctf_stream_destroy_listener {
 
 struct bt_ctf_stream {
 	struct bt_object base;
-	uint32_t id;
+	int64_t id;
 	struct bt_ctf_stream_class *stream_class;
 	GString *name;
 	struct bt_ctf_field *packet_header;
